requirements.txt
av==11.0.0
diffusers==0.35.0.dev0
diffusers.egg==info
easydict==1.13
imageio==2.37.0
matplotlib==3.10.5
numpy==2.3.2
opencv_python_headless==4.11.0.86
peft==0.17.0
Pillow==11.3.0
PIMS==0.7
setuptools==75.9.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
